fix: match dtype in BoundaryOffsetModule scatter_reduce_ for AMP compatibility
torch.full defaulted to float32 while edge_feat could be float16/bfloat16
under AMP autocast, causing RuntimeError in scatter_reduce_.

+++ project/models/heads.py
@@ -184,7 +184,8 @@ class BoundaryOffsetModule(nn.Module):
 
         # Max-pool over K neighbors per point
         pooled = torch.full(
-            (N, edge_feat.shape[1]), float("-inf"), device=edge_feat.device
+            (N, edge_feat.shape[1]), float("-inf"),
+            device=edge_feat.device, dtype=edge_feat.dtype,
         )
         pooled.scatter_reduce_(
             0, col.unsqueeze(-1).expand_as(edge_feat), edge_feat, reduce="amax"
